+++ utils/common/crc32.cpp
@@ -14,6 +14,7 @@ GNU General Public License for more details.
 */
 
 #include "cmdlib.h"
+#include <stdint.h>
 
 #define NUM_BYTES		256
 #define CRC32_INIT_VALUE	0xFFFFFFFFUL
@@ -137,7 +138,7 @@ JustAfew:
 	// the main loop is aligned and only has to worry about 8 byte at a time.
 	// The low-order two bits of pb and nBuffer in total control the
 	// upfront work.
-	nFront = ((uint)pb) & 3;
+	nFront = ((ptrdiff_t)pb) & 3;
 	nBuffer -= nFront;
 
 	switch( nFront )
